requirements.txt
aio-pika==9.5.5
aiormq==6.8.1
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
bidict==0.23.1
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.12
flake8==7.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
mccabe==0.7.0
motor==3.7.0
multidict==6.4.3
mypy==1.15.0
mypy-extensions==1.0.0
packaging==25.0
pamqp==3.3.0
passlib==1.7.4
pika==1.3.2
pluggy==1.5.0
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
pymongo==4.12.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
python-engineio==4.12.0
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
rsa==4.9
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.0
